Project structure and basic functionality
Images are displayed responsively in up to 5 columns. Images
are scaled down to 200px wide as this is the narrowest width
occurring in the source data, so all images in a given column
are the same width without scaling any up and degrading quality.

Very wide images are clipped to ensure that they retain a minimum
height limit while remaining 200px wide.

Mouseover opens a slider with the text associated with each image.