Prevent background shifting when ImageView opens
Setting the viewport div to `overflow: hidden` when opening
ImageView caused the page to shift because the scrollbar
disappeared.

Leaving `overflow: auto` all the time prevents this, although
it becomes possible to scroll the background while ImageView
is open using a middle mouse click.

The shifting background is much more noticeable while the
middle-mouse scrolling doesn't really cause a problem anyway,
so this commit removes the conditional `overflow: hidden`.