Add small transition delay to focus/hover dropdown
With no delay, moving the mouse over the page caused a lot of
movement as the text dropdown hover effect would begin to trigger
on each thumbnail passed over.

Adding a small delay makes the page much more stable while
still enabling the text to open quickly when the mouse stops
on a thumbnail.